backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.3
python-multipart==0.0.17
reportlab==4.2.5
pillow==11.0.0
python-dotenv==1.0.1
aiofiles==24.1.0
